requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
argon2-cffi==23.1.0
websockets==13.0
